implement reading crash reprots and directory configuration
