+++ .github/workflows/license.yaml
@@ -9,6 +9,6 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v3
-    - uses: EmbarkStudios/cargo-deny-action@v2.0.1
+    - uses: EmbarkStudios/cargo-deny-action@v2.0.6
       with:
         command: check licenses
