+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
   "bd-buffer",
   "bd-client-common",
   "bd-client-stats",
+  "bd-crash-handler",
   "bd-client-stats-store",
   "bd-completion",
   "bd-device",
@@ -75,11 +76,7 @@ http-body       = "1.0.1"
 http-body-util  = "0.1.2"
 hyper           = "1.6.0"
 
-hyper-rustls = { version = "0.27.5", default-features = false, features = [
-  "http1",
-  "http2",
-  "webpki-tokio",
-] }
+hyper-rustls = { version = "0.27.5", default-features = false, features = ["http1", "http2", "webpki-tokio"] }
 
 hyper-util            = { version = "0.1.10", features = ["client", "client-legacy"] }
 intrusive-collections = "0.9.7"
