+++ bd-logger/Cargo.toml
@@ -18,6 +18,7 @@ bd-client-common.path        = "../bd-client-common"
 bd-client-stats.path         = "../bd-client-stats"
 bd-client-stats-store.path   = "../bd-client-stats-store"
 bd-completion.path           = "../bd-completion"
+bd-crash-handler.path        = "../bd-crash-handler"
 bd-device.path               = "../bd-device"
 bd-events.path               = "../bd-events"
 bd-internal-logging.path     = "../bd-internal-logging"
