+++ bd-workflows/src/engine_test.rs
@@ -1263,7 +1263,7 @@ async fn persistence_to_disk_is_rate_limited() {
   engine_assert_active_run_traversals!(other_workflows_engine; 0 => 0; "B", "D");
 
   // Advance clock to allow rate limiting to kick in.
-  let elapsed = workflows_engine
+  let elapsed = *workflows_engine
     .state_store
     .persistence_write_interval_flag
     .read()
